requirements.txt
absl-py==2.4.0
altair==4.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
attrs==26.1.0
audio-recorder-streamlit==0.0.10
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
colorama==0.4.6
cryptography==46.0.6
distro==1.9.0
entrypoints==0.4
fastapi==0.135.3
flatbuffers==25.12.19
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
google-genai==1.69.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.73.1
grpcio==1.78.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
importlib-metadata==6.11.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.14.0
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
narwhals==2.18.1
numpy==2.4.3
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==2.3.3
pillow==10.4.0
proto-plus==1.27.2
protobuf==7.34.1
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2026.1.post1
referencing==0.37.0
requests==2.33.0
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SpeechRecognition==3.15.2
starlette==1.0.0
streamlit==1.38.0
streamlit_mic_recorder==0.0.8
tenacity==8.5.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.21.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
toml==0.10.2
toolz==1.1.0
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.43.0
validators==0.35.0
watchdog==4.0.2
websockets==16.0
Werkzeug==3.1.7
wrapt==1.14.2
zipp==3.23.0
